php7 not-well formed fix

+++ lib/eventum/class.misc.php
@@ -221,7 +221,7 @@ class Misc
      */
     public static function return_bytes($val)
     {
-        $val = trim($val);
+        $val = (int)trim($val);
         $last = strtolower($val[strlen($val) - 1]);
         switch ($last) {
             // The 'G' modifier is available since PHP 5.1.0
